update reccieve transaction api & send token api

+++ src/main/java/com/xavier/domain/AdminDomain.java
@@ -1,5 +1,6 @@
 package com.xavier.domain;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 
@@ -16,7 +17,7 @@ public class AdminDomain {
 	int id;
 	String fileName;
 	String address;
-	BigInteger ether;
+	BigDecimal ether;
 	String password;
 	Date dateCreated;
 	BigInteger token;
@@ -44,10 +45,10 @@ public class AdminDomain {
 	public void setAddress(String address) {
 		this.address = address;
 	}
-	public BigInteger getEther() {
+	public BigDecimal getEther() {
 		return ether;
 	}
-	public void setEther(BigInteger ether) {
+	public void setEther(BigDecimal ether) {
 		this.ether = ether;
 	}
 	public String getPassword() {
